+++ net/net_ykc_monitor/src/ykc_monitor_message_receive.c
@@ -204,7 +204,7 @@ static void ykc_monitor_set_recv_card_whitelists_data(uint8_t index, uint8_t sta
  * 函数名      ykc_monitor_get_recv_card_whitelists_count
 * 功能           根据信息项的指令码获取指定的接收信息项
  **********************************************/
-uint8_t ykc_monitor_get_recv_card_whitelists_count(uint8_t option)
+uint8_t ykc_monitor_get_recv_card_whitelists_count(uint32_t option)
 {
     if(option &NET_YKC_MONITOR_CARD_WHITELIST_QUERY){
         return s_ykc_monitor_card_query.count;
@@ -218,7 +218,7 @@ uint8_t ykc_monitor_get_recv_card_whitelists_count(uint8_t option)
  * 函数名      ykc_monitor_get_recv_card_whitelists_state
 * 功能           根据信息项的指令码获取指定的接收信息项
  **********************************************/
-uint32_t ykc_monitor_get_recv_card_whitelists_state(uint8_t option)
+uint32_t ykc_monitor_get_recv_card_whitelists_state(uint32_t option)
 {
     if(option &NET_YKC_MONITOR_CARD_WHITELIST_QUERY){
         return s_ykc_monitor_card_query.state;
@@ -232,7 +232,7 @@ uint32_t ykc_monitor_get_recv_card_whitelists_state(uint8_t option)
  * 函数名      ykc_monitor_clear_recv_card_whitelists_state
 * 功能           根据信息项的指令码清除指定的接收信息项
  **********************************************/
-void ykc_monitor_clear_recv_card_whitelists_state(uint8_t option)
+void ykc_monitor_clear_recv_card_whitelists_state(uint32_t option)
 {
     if(option &NET_YKC_MONITOR_CARD_WHITELIST_QUERY){
         s_ykc_monitor_card_query.state = 0x00;
